Add starter code for assignment
This change comprises starter code for the assignment -- i.e., the
declaration of functions that must be implemented in the functions
module.

Unit tests for these functions are also provided in the tests
directory. The unit tests use tox and pytest for automated testing. In
addition, stylstic checks are performed using Flake8.